Revert "Pro 5159 extend caveats"
