+++ ccdImports/configFiles/CCD_Probate_Caveat/AuthorisationCaseEvent.json
@@ -31,7 +31,6 @@
   {"CaseTypeID": "Caveat", "CaseEventID": "withdrawCaveatForAwaitingCaveatResolution", "UserRole": "caseworker-probate-issuer", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseEventID": "caveatExpiredForAwaitingCaveatResolution", "UserRole": "caseworker-probate-issuer", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseEventID": "warningRequested", "UserRole": "caseworker-probate-issuer", "CRUD": "CRUD"},
-  {"CaseTypeID": "Caveat", "CaseEventID": "extendCaveat", "UserRole": "caseworker-probate-issuer", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseEventID": "amendCaveatForWarningValidation", "UserRole": "caseworker-probate-issuer", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseEventID": "uploadDocumentForWarningValidation", "UserRole": "caseworker-probate-issuer", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseEventID": "emailCaveatorForWarningValidation", "UserRole": "caseworker-probate-issuer", "CRUD": "CRUD"},
@@ -107,7 +106,6 @@
   {"CaseTypeID": "Caveat", "CaseEventID": "withdrawCaveatForAwaitingCaveatResolution", "UserRole": "caseworker-probate-caseadmin", "CRUD": "R"},
   {"CaseTypeID": "Caveat", "CaseEventID": "caveatExpiredForAwaitingCaveatResolution", "UserRole": "caseworker-probate-caseadmin", "CRUD": "R"},
   {"CaseTypeID": "Caveat", "CaseEventID": "warningRequested", "UserRole": "caseworker-probate-caseadmin", "CRUD": "R"},
-  {"CaseTypeID": "Caveat", "CaseEventID": "extendCaveat", "UserRole": "caseworker-probate-caseadmin", "CRUD": "R"},
   {"CaseTypeID": "Caveat", "CaseEventID": "amendCaveatForWarningValidation", "UserRole": "caseworker-probate-caseadmin", "CRUD": "R"},
   {"CaseTypeID": "Caveat", "CaseEventID": "uploadDocumentForWarningValidation", "UserRole": "caseworker-probate-caseadmin", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseEventID": "emailCaveatorForWarningValidation", "UserRole": "caseworker-probate-caseadmin", "CRUD": "CRUD"},
@@ -178,7 +176,6 @@
   {"CaseTypeID": "Caveat", "CaseEventID": "withdrawCaveatForAwaitingCaveatResolution", "UserRole": "caseworker-probate-caseofficer", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseEventID": "caveatExpiredForAwaitingCaveatResolution", "UserRole": "caseworker-probate-caseofficer", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseEventID": "warningRequested", "UserRole": "caseworker-probate-caseofficer", "CRUD": "CRUD"},
-  {"CaseTypeID": "Caveat", "CaseEventID": "extendCaveat", "UserRole": "caseworker-probate-caseofficer", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseEventID": "amendCaveatForWarningValidation", "UserRole": "caseworker-probate-caseofficer", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseEventID": "uploadDocumentForWarningValidation", "UserRole": "caseworker-probate-caseofficer", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseEventID": "emailCaveatorForWarningValidation", "UserRole": "caseworker-probate-caseofficer", "CRUD": "CRUD"},
@@ -252,7 +249,6 @@
   {"CaseTypeID": "Caveat", "CaseEventID": "withdrawCaveatForAwaitingCaveatResolution", "UserRole": "caseworker-probate-registrar", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseEventID": "caveatExpiredForAwaitingCaveatResolution", "UserRole": "caseworker-probate-registrar", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseEventID": "warningRequested", "UserRole": "caseworker-probate-registrar", "CRUD": "CRUD"},
-  {"CaseTypeID": "Caveat", "CaseEventID": "extendCaveat", "UserRole": "caseworker-probate-registrar", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseEventID": "amendCaveatForWarningValidation", "UserRole": "caseworker-probate-registrar", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseEventID": "uploadDocumentForWarningValidation", "UserRole": "caseworker-probate-registrar", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseEventID": "emailCaveatorForWarningValidation", "UserRole": "caseworker-probate-registrar", "CRUD": "CRUD"},
@@ -327,7 +323,6 @@
   {"CaseTypeID": "Caveat", "CaseEventID": "caveatExpiredForAwaitingCaveatResolution", "UserRole": "caseworker-probate-superuser", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseEventID": "findMatchedCaseForImport", "UserRole": "caseworker-probate-superuser", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseEventID": "warningRequested", "UserRole": "caseworker-probate-superuser", "CRUD": "CRUD"},
-  {"CaseTypeID": "Caveat", "CaseEventID": "extendCaveat", "UserRole": "caseworker-probate-superuser", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseEventID": "amendCaveatForWarningValidation", "UserRole": "caseworker-probate-superuser", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseEventID": "uploadDocumentForWarningValidation", "UserRole": "caseworker-probate-superuser", "CRUD": "CRUD"},
   {"CaseTypeID": "Caveat", "CaseEventID": "emailCaveatorForWarningValidation", "UserRole": "caseworker-probate-superuser", "CRUD": "CRUD"},
