+++ compose/dm-store.yml
@@ -17,7 +17,7 @@ services:
 
   ccd-data-store-api:
     environment:
-      CCD_DM_DOMAIN: http://localhost:5006
+      CCD_DM_DOMAIN: http://dm-store:8080
 
   service-auth-provider-api:
     environment:
