+++ dependency-check-suppressions.xml
@@ -423,88 +423,4 @@ file name: spring-jcl-5.1.5.RELEASE.jar
         <gav regex="true">^org\.springframework:spring-jcl:.*$</gav>
         <cve>CVE-2020-5398</cve>
     </suppress>
-    <suppress>
-        <notes><![CDATA[
-   file name: spring-webmvc-5.1.5.RELEASE.jar
-   ]]></notes>
-        <gav regex="true">^org\.springframework:spring-webmvc:.*$</gav>
-        <cve>CVE-2020-5398</cve>
-    </suppress>
-    <suppress>
-        <notes><![CDATA[
-   file name: spring-oxm-5.1.5.RELEASE.jar
-   ]]></notes>
-        <gav regex="true">^org\.springframework:spring-oxm:.*$</gav>
-        <cve>CVE-2020-5398</cve>
-    </suppress>
-    <suppress>
-        <notes><![CDATA[
-   file name: spring-context-5.1.5.RELEASE.jar
-   ]]></notes>
-        <gav regex="true">^org\.springframework:spring-context:.*$</gav>
-        <cve>CVE-2020-5398</cve>
-    </suppress>
-    <suppress>
-        <notes><![CDATA[
-   file name: spring-aop-5.1.5.RELEASE.jar
-   ]]></notes>
-        <gav regex="true">^org\.springframework:spring-aop:.*$</gav>
-        <cve>CVE-2020-5398</cve>
-    </suppress>
-    <suppress>
-        <notes><![CDATA[
-   file name: spring-aspects-5.1.5.RELEASE.jar
-   ]]></notes>
-        <gav regex="true">^org\.springframework:spring-aspects:.*$</gav>
-        <cve>CVE-2020-5398</cve>
-    </suppress>
-    <suppress>
-        <notes><![CDATA[
-   file name: spring-orm-5.1.5.RELEASE.jar
-   ]]></notes>
-        <gav regex="true">^org\.springframework:spring-orm:.*$</gav>
-        <cve>CVE-2020-5398</cve>
-    </suppress>
-    <suppress>
-        <notes><![CDATA[
-   file name: spring-jdbc-5.1.5.RELEASE.jar
-   ]]></notes>
-        <gav regex="true">^org\.springframework:spring-jdbc:.*$</gav>
-        <cve>CVE-2020-5398</cve>
-    </suppress>
-    <suppress>
-        <notes><![CDATA[
-   file name: spring-tx-5.1.5.RELEASE.jar
-   ]]></notes>
-        <gav regex="true">^org\.springframework:spring-tx:.*$</gav>
-        <cve>CVE-2020-5398</cve>
-    </suppress>
-    <suppress>
-        <notes><![CDATA[
-   file name: spring-beans-5.1.5.RELEASE.jar
-   ]]></notes>
-        <gav regex="true">^org\.springframework:spring-beans:.*$</gav>
-        <cve>CVE-2020-5398</cve>
-    </suppress>
-    <suppress>
-        <notes><![CDATA[
-   file name: spring-expression-5.1.5.RELEASE.jar
-   ]]></notes>
-        <gav regex="true">^org\.springframework:spring-expression:.*$</gav>
-        <cve>CVE-2020-5398</cve>
-    </suppress>
-    <suppress>
-        <notes><![CDATA[
-   file name: spring-core-5.1.5.RELEASE.jar
-   ]]></notes>
-        <gav regex="true">^org\.springframework:spring-core:.*$</gav>
-        <cve>CVE-2020-5398</cve>
-    </suppress>
-    <suppress>
-        <notes><![CDATA[
-   file name: spring-jcl-5.1.5.RELEASE.jar
-   ]]></notes>
-        <gav regex="true">^org\.springframework:spring-jcl:.*$</gav>
-        <cve>CVE-2020-5398</cve>
-    </suppress>
 </suppressions>
