+++ src/main/java/uk/gov/hmcts/probate/transformer/CCDDataTransformer.java
@@ -147,6 +147,38 @@ public class CCDDataTransformer {
         }
     }
 
+    public CaveatData transformCaveats(CaveatCallbackRequest callbackRequest) {
+
+        return buildCCDDataCaveats(callbackRequest);
+    }
+
+    private CaveatData buildCCDDataCaveats(CaveatCallbackRequest callbackRequest) {
+        CaveatData caseData = callbackRequest.getCaseDetails().getData();
+
+        return CaveatData.builder()
+                .caveatorEmailAddress(notNullWrapper(caseData.getCaveatorEmailAddress()))
+                .build();
+    }
+
+    public CaveatData transformSolsCaveats(CaveatCallbackRequest caveatCallbackRequest) {
+
+        return buildCCDDataSolsCaveats(caveatCallbackRequest);
+    }
+
+    private CaveatData buildCCDDataSolsCaveats(CaveatCallbackRequest caveatCallbackRequest) {
+        CaveatData caveatData = caveatCallbackRequest.getCaseDetails().getData();
+
+        return CaveatData.builder()
+                .registryLocation(notNullWrapper(caveatData.getRegistryLocation()))
+                .solsSolicitorAppReference(notNullWrapper(caveatData.getSolsSolicitorAppReference()))
+                .applicationSubmittedDate(getCaseSubmissionDate(caveatCallbackRequest.getCaseDetails()
+                        .getLastModified()))
+                .caveatorEmailAddress(notNullWrapper(caveatData.getCaveatorEmailAddress()))
+                .solsPaymentMethods(notNullWrapper(caveatData.getSolsPaymentMethods()))
+                .solsFeeAccountNumber(notNullWrapper(caveatData.getSolsFeeAccountNumber()))
+                .build();
+    }
+
     public CCDData transformBulkPrint(String letterId) {
 
         return buildCCDDataBulkPrint(letterId);
