+++ src/main/java/uk/gov/hmcts/probate/validator/CaveatsEmailAddressNotificationValidationRule.java
@@ -1,5 +1,5 @@
 package uk.gov.hmcts.probate.validator;
 
-public interface CaveatsEmailAddressNotificationValidationRule extends CaveatsEmailValidationRule {
+public interface CaveatsEmailAddressNotificationValidationRule extends ValidationRuleCaveats {
 }
