--- src/main/java/uk/gov/hmcts/probate/validator/CaveatsEmailValidationRule.java
@@ -1,5 +1,0 @@
-package uk.gov.hmcts.probate.validator;
-
-public interface CaveatsEmailValidationRule extends ValidationRuleCaveats {
-}
-
