--- src/main/java/uk/gov/hmcts/probate/validator/CaveatsExpiryValidationRule.java
@@ -1,4 +1,0 @@
-package uk.gov.hmcts.probate.validator;
-
-public interface CaveatsExpiryValidationRule extends ValidationRuleCaveats {
-}
