+++ src/main/resources/ValidationMessages.properties
@@ -68,9 +68,6 @@ questionCanNotBeNo=Ensure all checks have been completed, cancel to return to th
 emailNotProvidedPA=There is no email address for this applicant. To continue the application, go back and select no to sending an email.
 emailNotProvidedSOLS=There is no email address for this solicitor. To continue the application, go back and select no to sending an email.
 emailNotProvidedCaveats=There is no email address for this caveator. Add an email address or contact them by post.
-caveatExpiryMoreThan1MonthAway=Cannot extend a caveat that is more than 1 month from expiry.
-caveatExpiryInThePast=Cannot extend an already expired caveat.
-caveatExpiryCannotBeInPast=Caveat expiry date cannot be set to be in the past.
 notifyApplicantNoEmailSOLS=There is no email address for this solicitor. Add an email address or contact them by post.
 notifyApplicantNoEmailPA=There is no email address for this applicant. Add an email address or contact them by post.
 legacyCaseSubmissionNotAllowed=Saving the legacy search case is not permitted. Please return to the case details to begin a new search
